# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/devices/register-token": {
      "post": {
        "operationId": "NotificationsController_registerDeviceToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceTokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token or platform payload"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register or refresh a device push token",
        "tags": [
          "Notifications",
          "Devices"
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_listNotifications",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "UNREAD",
                "READ"
              ]
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or status filter"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List current user notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markNotificationRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Only notification owner can mark as read"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark one notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/cycles/{cycleId}/payout": {
      "post": {
        "operationId": "PayoutsController_createPayout",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cycle closed, payout already exists, or invalid proof scope"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Cycle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create payout for an open cycle",
        "tags": [
          "Payouts"
        ]
      },
      "get": {
        "operationId": "PayoutsController_getCyclePayout",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          },
          "404": {
            "description": "Cycle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payout for cycle if available",
        "tags": [
          "Payouts",
          "Cycles"
        ]
      }
    },
    "/payouts/{id}/confirm": {
      "patch": {
        "operationId": "PayoutsController_confirmPayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPayoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Payout state or strict payout checks failed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Payout not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a pending payout",
        "tags": [
          "Payouts"
        ]
      }
    },
    "/cycles/{cycleId}/close": {
      "post": {
        "operationId": "PayoutsController_closeCycle",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cycle is closed or payout is not confirmed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Cycle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close cycle after payout is confirmed",
        "tags": [
          "Payouts"
        ]
      }
    }
  },
  "info": {
    "title": "Equb API",
    "description": "Equb backend API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDeviceTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID",
              "WEB"
            ]
          }
        },
        "required": [
          "token",
          "platform"
        ]
      },
      "DeviceTokenResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID",
              "WEB"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "token",
          "platform",
          "isActive",
          "lastSeenAt",
          "createdAt"
        ]
      },
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "groupId": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "MEMBER_JOINED",
              "CONTRIBUTION_SUBMITTED",
              "CONTRIBUTION_CONFIRMED",
              "CONTRIBUTION_REJECTED",
              "PAYOUT_CONFIRMED",
              "DUE_REMINDER"
            ]
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "dataJson": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "enum": [
              "UNREAD",
              "READ"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "readAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "body",
          "status",
          "createdAt"
        ]
      },
      "NotificationListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ]
      },
...
      "CreatePayoutDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 500
          },
          "proofFileKey": {
            "type": "string"
          },
          "paymentRef": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "PayoutUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "nullable": true
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "phone"
        ]
      },
      "PayoutResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "toUserId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED"
            ]
          },
          "proofFileKey": {
            "type": "object",
            "nullable": true
          },
          "paymentRef": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "confirmedByUserId": {
            "type": "object",
            "nullable": true
          },
          "confirmedAt": {
            "type": "object",
            "nullable": true
          },
          "toUser": {
            "$ref": "#/components/schemas/PayoutUserResponseDto"
          }
        },
        "required": [
          "id",
          "groupId",
          "cycleId",
          "toUserId",
          "amount",
          "status",
          "createdByUserId",
          "createdAt",
          "toUser"
        ]
      },
      "ConfirmPayoutDto": {
        "type": "object",
        "properties": {
          "proofFileKey": {
            "type": "string"
          },
          "paymentRef": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      }
    }
  }
}